fix missing DINO v3 offline error

+++ embeddings/dinov2.py
@@ -52,9 +52,9 @@ class DINOv2Embedding(EmbeddingModel):
 
     def encode(self, image: np.ndarray, mask: np.ndarray) -> np.ndarray:
         crop = padded_square_crop(image, mask, size=224)
-        crop = crop.astype(np.float32) / 255.0
-        mean = np.array([0.485, 0.456, 0.406])
-        std = np.array([0.229, 0.224, 0.225])
+        crop = (crop / 255.0).astype(np.float32)
+        mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
         crop = (crop - mean) / std
         tensor = torch.from_numpy(crop).permute(2, 0, 1).unsqueeze(0).to(self.device)
